# Prevent extra string allocations when no 'rel' arg passed