@natematykiewicz 2017-11-02
Looks like the entire purpose of the lstrip was to account for there not being a rel passed in. With this change, it seems like you could remove the lstrip, which would remove 1 more string allocation.