@oniofchaos 2017-11-02
You could pass in something like `rel: nil` or `rel: ''`, which would result in `'  nofollow'` if I took out the `lstrip`