@natematykiewicz 2017-11-02
If you changed your `if` to a `present?` check instead of `has_key?`, you could do it.

Or you could just verify that it’s not nil, and if someone passes in “”, then they can live with the leading space in the attribute name.

Honestly, I think you should do one of those two over the `has_key?`.

It just feels to me like it’s an easy win, and if people are doing weird things, they deserve weird html. Instead of penalizing everyone else for a few people’s weird use of a `rel` attribute.