@oniofchaos 2017-11-02
Thanks for the feedback. Ended up going with a `blank?` check. See https://github.com/rails/rails/pull/30997/commits/d404e2cbe9e078d286c7acb13fd18fb3c04a3f4f for the benchmark details, but it's performant in all of the scenarios I could think of and handles the edge case of passing in something like `rel: ''`.

Thoughts on this solution?