@natematykiewicz 2017-11-03
Nice! I’ve never seen that `.()` notation. Is it different from `[]`?